chore: e2e: create a helper.mjs
