+++ e2e/helper.mjs
@@ -1,0 +1,36 @@
+import { test, expect } from '@playwright/test';
+import * as fs from 'fs';
+import { fileURLToPath } from 'url';
+
+/**
+ * Read the current file and fetch all dynamic import dependencies using regex.
+ * TODO use AST instead of regex
+ */
+const fetchAllTestsDependencies = async (dirname) => {
+  const currentFile = (await fs.promises.readFile(dirname)).toString();
+  const importRegexp = /import\((.+)\)/gm;
+  const matches = [...currentFile.matchAll(importRegexp)];
+  const dependencies = matches.map(arrayItem => {
+    const dependency = arrayItem[1];
+    return dependency.replace(/['"]+/g, '');
+  });  
+  return dependencies;
+};
+
+/**
+ * Generates importMap with all the dependencies in the tests in this file and returns importMap as a string
+ */
+const getImportMap = async (resolve, dirname) => {
+  const dependencies = await fetchAllTestsDependencies(dirname);
+  const importMapObject = {
+    imports: {},
+  };
+  dependencies.forEach(dependency => {
+    importMapObject.imports[dependency] = resolve(dependency).split(process.env.PWD)[1];
+  });
+  return JSON.stringify(importMapObject);
+};
+
+export const goToPage = async (page, resolve, dirname) => {
+  await page.goto(`http://localhost:8005/?importMap=${await getImportMap(resolve, dirname)}`);
+};
