+++ packages/ui/components/combobox/test/lion-combobox.e2e.spec.js
@@ -1,47 +1,15 @@
 import { test, expect } from '@playwright/test';
-import * as fs from 'fs';
 import { fileURLToPath } from 'url';
 const __dirname = fileURLToPath(import.meta.url);
-
-/**
- * Read the current file and fetch all dynamic import dependencies using regex.
- * TODO use AST instead of regex
- */
-const fetchAllTestsDependencies = async () => {
-  const currentFile = (await fs.promises.readFile(__dirname)).toString();
-  const importRegexp = /import\((.+)\)/gm;
-  const matches = [...currentFile.matchAll(importRegexp)];
-  const dependencies = matches.map(arrayItem => {
-    const dependency = arrayItem[1];
-    return dependency.replace(/['"]+/g, '');
-  });
-  return dependencies;
-};
-
-/**
- * Generates importMap with all the dependencies in the tests in this file and returns importMap as a string
- */
-const getImportMap = async () => {
-  const dependencies = await fetchAllTestsDependencies();
-  const importMapObject = {
-    imports: {},
-  };
-  dependencies.forEach(dependency => {
-    importMapObject.imports[dependency] = import.meta.resolve(dependency).split(process.env.PWD)[1];
-  });
-  return JSON.stringify(importMapObject);
-};
-
-const goToPage = async page => {
-  await page.goto(`http://localhost:8005/?importMap=${await getImportMap()}`);
-};
+import { goToPage } from '../../../../../e2e/helper.mjs';
 
 test.describe('lion-combobox', () => {
   // TODO Fix the lion source code
   test.skip('Combobox does not flash the menu when _showOverlayCondition returns "false"', async ({
     page,
   }, testInfo) => {
-    await goToPage(page);
+    await goToPage(page, import.meta.resolve, __dirname);
+    //await goToPage(page);
     await page.evaluate(async () => {
       const { html, render } = await import('lit');
       await import('@lion/ui/define/lion-combobox.js');
@@ -124,49 +92,9 @@ test.describe('lion-combobox', () => {
 
   test("doesn't select any similar options after using delete when requireOptionMatch is false", async ({
     page,
-  }, testInfo) => {
-    await goToPage(page);
-    await page.evaluate(async () => {
-      const { html, render } = await import('lit');
-      const { Required } = await import('@lion/ui/form-core.js');
-      await import('@lion/ui/define/lion-combobox.js');
-      await import('@lion/ui/define/lion-option.js');
-      const { loadDefaultFeedbackMessages } = await import('@lion/ui/validate-messages.js');
-      loadDefaultFeedbackMessages();
-
-      const template = () => html`
-        <lion-combobox name="foo" .validators=${[new Required()]}>
-          <lion-option .choiceValue="${'Artichoke'}">Artichoke</lion-option>
-          <lion-option .choiceValue="${'Chard'}">Chard</lion-option>
-          <lion-option .choiceValue="${'Chicory'}">Chicory</lion-option>
-          <lion-option .choiceValue="${'Victoria Plum'}">Victoria Plum</lion-option>
-        </lion-combobox>
-      `;
-
-      render(template(), document.body);
-    });
-
-    await page.evaluate(() => {
-      const comboboxEl = document.querySelector('lion-combobox');
-      comboboxEl.requireOptionMatch = false;
-    });
-
-    const combobox = await page.locator('lion-combobox');
-    const input = await page.locator('css=input');
-    await input.focus();
-    await page.keyboard.type('a');
-    await page.keyboard.type('r');
-    await page.keyboard.type('t');
-    await page.keyboard.press('Backspace');
-    await page.keyboard.press('Enter');
-
-    expect(await combobox.evaluate(el => el.checkedIndex)).toEqual(-1);
-    expect(await combobox.evaluate(el => el.modelValue)).toEqual('Art');
-    expect(await combobox.evaluate(el => el.value)).toEqual('Art');
-  });
-
-  test('Test inline script execution', async ({ page }, testInfo) => {
-    await goToPage(page);
+  }) => {
+    //await goToPage(page);
+    await goToPage(page, import.meta.resolve, __dirname);
     await page.evaluate(async () => {
       const { html, render } = await import('lit');
       const { Required } = await import('@lion/ui/form-core.js');
@@ -209,7 +137,8 @@ test.describe('lion-combobox', () => {
   test('allows new options when multi-choice when requireOptionMatch=false and autocomplete="both", when deleting autocomplete values using Backspace', async ({
     page,
   }, testInfo) => {
-    await goToPage(page);
+    await goToPage(page, import.meta.resolve, __dirname);
+    //await goToPage(page);
     await page.evaluate(async () => {
       const { html, render } = await import('lit');
       await import('@lion/ui/define/lion-combobox.js');
@@ -244,7 +173,8 @@ test.describe('lion-combobox', () => {
   test('hides listbox on click/enter (when multiple-choice is false)', async ({
     page,
   }, testInfo) => {
-    await goToPage(page);
+    await goToPage(page, import.meta.resolve, __dirname);
+    //await goToPage(page);
     await page.evaluate(async () => {
       const { html, render } = await import('lit');
       await import('@lion/ui/define/lion-combobox.js');
